+++ lib/actions/backend/authBackendDashboard.action.php
@@ -1,0 +1,14 @@
+<?php
+
+class authBackendDashboardAction extends authBackendDomainSettingsAction
+{
+    protected function getUrlSegment(): string
+    {
+        return '';
+    }
+
+    protected function getTemplateName(): string
+    {
+        return 'BackendDashboard';
+    }
+}